extract Level Details (description & discussion) into separate component
