+++ src/pages/LevelSelectPage/LevelSelectPage.module.css
@@ -1,30 +1,3 @@
-.unselected{
-  width:50%;
-  font-size:16pt; 
-  font-weight:bold;
-  background-color: #3E3E3E;
-  color: #B3B3B3;
-}
-
-.selected{
-  width:50%;
-  font-size:16pt; 
-  font-weight:bold;
-  background-color: #7A4CC5;
-}
-
-.unselected:hover,
-.selected:hover {
-  border-color: #fff;
-}
-
-.unselected:focus,
-.unselected:focus-visible,
-.selected:focus,
-.selected:focus-visible {
-  outline: none;
-}
-
 .mainGrid{
   position: absolute;
   top: 100px;
